Move reload analyzers to server
This is need for Synonyms Management API project to be able
to auto-reload analyzers with synonyms on synonms change.
