+++ server/src/main/java/org/elasticsearch/action/admin/indices/analyze/ReloadAnalyzerAction.java
@@ -1,10 +1,12 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
-package org.elasticsearch.xpack.core.action;
+
+package org.elasticsearch.action.admin.indices.analyze;
 
 import org.elasticsearch.action.ActionType;
